+++ test/Qowaiv.UnitTests/Globalization/CultureInfoScopeTest.cs
@@ -38,15 +38,5 @@ namespace Qowaiv.UnitTests.Globalization
             Assert.AreEqual(current, CultureInfo.CurrentCulture);
             Assert.AreEqual(currentUI, CultureInfo.CurrentUICulture);
         }
-
-        [Test]
-        public void DebuggerDisplay_WithResult()
-        {
-            using (new CultureInfoScope("en-NL", "en-US"))
-            {
-                using var scope = new CultureInfoScope("es-ES", "fr-FR");
-                DebuggerDisplayAssert.HasResult("CultureInfoScope: [es-ES/fr-FR], Previous: [en-NL/en-US]", scope);
-            }
-        }
     }
 }
